requirements.txt
asgiref==3.7.2
dj-database-url==2.1.0
Django==4.2.7
django-cors-headers==4.3.1
django-environ==0.11.2
djangorestframework==3.14.0
gunicorn==21.2.0
joblib==1.3.2
numpy==1.26.2
packaging==23.2
psycopg2-binary==2.9.9
pytz==2023.3.post1
scikit-learn==1.2.2
scipy==1.11.4
sqlparse==0.4.4
threadpoolctl==3.2.0
typing_extensions==4.8.0
tzdata==2023.3
